Setup recursive file loading and made an attempt to organize files.
