+++ commands/points/points.js
@@ -1,8 +1,8 @@
 const Discord = require("discord.js");
 const mongoose = require("mongoose");
-const ids = require(`../config.json`);
-const secrets = require(`../secrets.json`);
-const pointsAdd = require("../models/addPoints.js");
+const ids = require(`../../config.json`);
+const secrets = require(`../../secrets.json`);
+const pointsAdd = require("../../models/addPoints.js");
 mongoose.connect(secrets.Mongo, {
     useUnifiedTopology: true,
     useNewUrlParser: true,
