+++ commands/weekly/weekly.js
@@ -1,9 +1,9 @@
 const Discord = require("discord.js");
 const mongoose = require("mongoose");
-const ids = require(`../config.json`);
-const updateWeekly = require("../models/weeklyUpdate.js");
-const secrets = require(`../secrets.json`);
-var tools = require(`../tools/functions.js`);
+const ids = require(`../../config.json`);
+const updateWeekly = require("../../models/weeklyUpdate.js");
+const secrets = require(`../../secrets.json`);
+var tools = require(`../../tools/functions.js`);
 mongoose.connect(secrets.Mongo, {
     useUnifiedTopology: true,
     useNewUrlParser: true,
