+++ commands/weekly/weeklyCron.js
@@ -1,7 +1,7 @@
-const ids = require(`../config.json`);
-const secrets = require(`../secrets.json`);
+const ids = require(`../../config.json`);
+const secrets = require(`../../secrets.json`);
 const mongoose = require("mongoose");
-const updateWeekly = require("../models/weeklyUpdate.js");
+const updateWeekly = require("../../models/weeklyUpdate.js");
 const { Channel } = require("diagnostics_channel");
 mongoose.connect(secrets.Mongo, {
     useUnifiedTopology: true,
